pyproject.toml
[project]
name = "gemma-api-repro"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.55.0",
    "python-dotenv>=1.0.1",
    "datasets>=4.8.4",
    "matplotlib>=3.9.0",
    "numpy>=2.0.0",
    "pandas>=2.2.0",
]
